Extract manage-files flow into WindowFileManagerMixin
Continue decomposing window.py: move the manage-files entry action and its
contiguous flow — first-run / operation-mode / backup prompts, the spawn
placeholder tab, and file-manager tab registration — verbatim into
sshpilot/window_file_manager.py as WindowFileManagerMixin, inherited by
MainWindow. window.py shrinks ~770 lines. Pure code move; the module-level
names the block relied on (Capability/capabilities_for, the
file_manager_integration helpers, connection host/alias, sftp_utils,
datetime/Path/Optional) are imported into the mixin so behavior is unchanged.

Place the mixin before WindowActions so its on_manage_files_action wins, and
delete the old dead on_manage_files_action duplicate from WindowActions (the
window.py body method already shadowed it via MRO).

The synchronous file-manager embed teardown stays in window.py for now (it is
shared with tab close/shutdown).

Add tests/test_window_file_manager_mixin.py asserting the methods resolve to
the mixin module, the mixin precedes WindowActions in the MRO, and the dead
duplicate stays gone.
